# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: ERP Backend API (Phase 1 + Phase 2 Inc1 + Inc2 LV §9 + FIN-0 Finance Stubs)
  version: 1.7.0-password-reset
  description: |
    **PWA / Browser-Anbindung**
    - **CORS**: Umgebungsvariable `CORS_ORIGINS` — kommagetrennte, exakte Origins (z. B. `https://app.example.com,http://localhost:5173`). Leer oder unset: keine `Access-Control-*`-Header (gleiche Origin oder Server-zu-Server).
    - **Auth**: `Authorization: Bearer <signed-v1-token>` (siehe `bearerToken`); optionaler Header `X-Tenant-Id` muss zum Tenant im Token passen, sonst `TENANT_SCOPE_VIOLATION`. **Passwort-Login:** `POST /auth/login` mit JSON `{ tenantId, email, password }` — siehe `docs/authentication-login.md`. **Passwort vergessen (öffentlich):** `POST /auth/request-password-reset`, `POST /auth/confirm-password-reset` — nur mit Postgres; Link `#/password-reset?token=…` (Basis `ERP_PUBLIC_APP_BASE_URL`); Produktion: SMTP + Basis-URL Pflicht. **Benutzerverwaltung:** `GET|POST /users`, `PATCH /users/{userId}` inkl. optional **E-Mail** (nur Rolle **ADMIN**, nur mit Postgres) — Audit-Ereignisse `USER` / `USER_CREATED` / `USER_UPDATED`. **Betrieb:** `AUTH_TOKEN_SECRET` (min. 32 Zeichen bei `NODE_ENV=production`); ohne Secret kein Start — siehe Repo-`.env.example`. Nur lokaler Demo-Modus: `ERP_ALLOW_INSECURE_DEV_AUTH=1` und nicht `production` (niemals deployen).
...
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: signed-v1-token
  parameters:
    TenantScopeHeader:
      in: header
      name: x-tenant-id
      required: false
      schema:
        type: string
        format: uuid
      description: Optionaler Tenant-Scope, muss mit Token-Tenant identisch sein.
...
  schemas:
...
    Error:
      type: object
      required: [code, message, correlationId, retryable, blocking]
      properties:
        code:
          type: string
          description: DomainError-Code oder VALIDATION_FAILED (siehe docs/contracts/error-codes.json).
        message:
          type: string
        correlationId:
          type: string
          format: uuid
          description: Pro Antwort von handleError generiert (Phase 1).
        retryable:
          type: boolean
        blocking:
          type: boolean
        details:
          type: object
          additionalProperties: true
          description: Optional, z. B. validationErrors bei EXPORT_PREFLIGHT_FAILED.
...
    TenantUser:
      type: object
      required: [id, email, role, active, createdAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [ADMIN, BUCHHALTUNG, GESCHAEFTSFUEHRUNG, VERTRIEB, VIEWER]
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    TenantUserListResponse:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/TenantUser'
    CreateTenantUserRequest:
      type: object
      required: [email, password, role, reason]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
          maxLength: 2000
        role:
          type: string
          enum: [ADMIN, BUCHHALTUNG, GESCHAEFTSFUEHRUNG, VERTRIEB, VIEWER]
        reason:
          type: string
          minLength: 5
    PatchTenantUserRequest:
      type: object
      required: [reason]
      properties:
        role:
          type: string
          enum: [ADMIN, BUCHHALTUNG, GESCHAEFTSFUEHRUNG, VERTRIEB, VIEWER]
        active:
          type: boolean
        password:
          type: string
          minLength: 12
          maxLength: 2000
        email:
          type: string
          format: email
          description: Neue E-Mail (normalisiert wie bei Login); mindestens ein Feld neben reason erforderlich.
        reason:
          type: string
          minLength: 5
...
paths:
...
  /users:
    get:
      tags: [Users (Admin)]
      operationId: tenantUsersList
      summary: Benutzer des Mandanten auflisten
      description: Nur **ADMIN**. Ohne Postgres `503` / `USER_MANAGEMENT_REQUIRES_DB`.
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
      responses:
        '200':
          description: Liste ohne Passwort-Hashes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantUserListResponse'
        '400':
          description: VALIDATION_FAILED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: AUTH_ROLE_FORBIDDEN
        '401':
          description: UNAUTHORIZED
        '503':
          description: USER_MANAGEMENT_REQUIRES_DB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [Users (Admin)]
      operationId: tenantUserCreate
      summary: Benutzer anlegen
      description: Nur **ADMIN**; E-Mail mandantenweit eindeutig. Audit `USER_CREATED`.
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantUserRequest'
      responses:
        '201':
          description: Angelegter Benutzer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantUser'
        '400':
          description: VALIDATION_FAILED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: AUTH_ROLE_FORBIDDEN
        '401':
          description: UNAUTHORIZED
        '409':
          description: USER_EMAIL_CONFLICT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: USER_MANAGEMENT_REQUIRES_DB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userId}:
    patch:
      tags: [Users (Admin)]
      operationId: tenantUserPatch
      summary: Benutzer aktualisieren (Rolle, aktiv, Passwort)
      description: |
        Nur **ADMIN**. Mindestens eines von `role`, `active`, `password` erforderlich (plus `reason`).
        Schutz letzter aktiver Admin und kein Deaktivieren des eigenen Kontos. Audit `USER_UPDATED`.
      security:
        - bearerToken: []
      parameters:
        - $ref: '#/components/parameters/TenantScopeHeader'
        - in: path
          name: userId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchTenantUserRequest'
      responses:
        '200':
          description: Aktualisierter Benutzer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantUser'
        '400':
          description: VALIDATION_FAILED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: AUTH_ROLE_FORBIDDEN
        '401':
          description: UNAUTHORIZED
        '404':
          description: USER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: USER_EMAIL_CONFLICT | USER_LAST_ADMIN_FORBIDDEN | USER_SELF_DEACTIVATION_FORBIDDEN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: USER_MANAGEMENT_REQUIRES_DB
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...